requirements.txt
python-dotenv
requests
xmltodict
schedule
boto3